[DEV-0002] Added Income and Expenses
Able to input Annual Income and Expenses (as percentages of income) for
now so that you're able to configure how much expenses you'll be using
out of your income as a percentage. it is still in progress to change.
This is only placed as a placeholder for the eventual intended effect
for it to be calculated in statistics.

+++ BudgetingApplication/src/Statistics.java
@@ -15,9 +15,9 @@ public class Statistics {
 		currency = new DecimalFormat("##.##");
 	}
 	
-	public void inputIncome(double i)
+	public void inputIncome(int i)
 	{
-		income = i;
+		income = (double) i;
 	}
 	
 	public void inputExpenses(double e)
